third_party/WebKit/LayoutTests/fast/forms/color/input-color-onchange-event-expected.txt

Test if change event fires properly when color chooser changes. Bug 66848 
To manually test this, click on the input color element in the top left corner and change the value from the color chooser. See if the number of "value changed" messages matches the number of times you changed the color.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.selectColorInColorChooser({}, '#ff0000'); threw exception TypeError: Failed to execute 'selectColorInColorChooser' on 'Internals': parameter 1 is not of type 'Element'..
PASS internals.selectColorInColorChooser(document, '#ff0000'); threw exception TypeError: Failed to execute 'selectColorInColorChooser' on 'Internals': parameter 1 is not of type 'Element'..
value changed to #ff0000
PASS input.value is "#ff0000"
PASS successfullyParsed is true

TEST COMPLETE
